Disassemble DSP microcode

+++ libs/RVL_SDK/include/revolution/ax/DSPCode.h
@@ -8,7 +8,7 @@ extern "C" {
 extern u16 axDspInitVector;
 extern u16 axDspResumeVector;
 
-extern u8 axDspSlave[];
+extern u16 axDspSlave[];
 extern u16 axDspSlaveLength;
 
 #ifdef __cplusplus
